move activator template into tab label component
