+++ libs/design/tabs/examples/src/basic-tabs/basic-tabs.component.html
@@ -1,17 +1,18 @@
 <daff-tabs>
 	<daff-tab>
-		<ng-template daffTabLabel>
-			<fa-icon [icon]="faInfoCircle" daffPrefix></fa-icon> Tab 1
-		</ng-template>
+		<daff-tab-label>
+			<fa-icon [icon]="faInfoCircle" daffPrefix></fa-icon>
+			<div>Tab 1</div>
+		</daff-tab-label>
 		<daff-tab-panel>
 			Tab 1 Content
 		</daff-tab-panel>
 	</daff-tab>
 
 	<daff-tab>
-		<ng-template daffTabLabel>
+		<daff-tab-label>
 			<fa-icon [icon]="faInfoCircle" daffPrefix></fa-icon> Tab 2
-		</ng-template>
+		</daff-tab-label>
 		<daff-tab-panel>
 			Tab 2 Content
 		</daff-tab-panel>
