+++ libs/design/tabs/src/public_api.ts
@@ -2,4 +2,4 @@ export { DAFF_TABS_COMPONENTS } from './tabs';
 export { DaffTabsComponent } from './tabs/tabs.component';
 export { DaffTabPanelComponent } from './tabs/tab-panel/tab-panel.component';
 export { DaffTabComponent } from './tabs/tab/tab.component';
-export { DaffTabLabelDirective } from './tabs/tab-label/tab-label.directive';
+export { DaffTabLabelComponent } from './tabs/tab-label/tab-label.component';
