--- libs/design/tabs/src/tabs/tab-activator/tab-activator.component.html
@@ -1,6 +1,0 @@
-<ng-container *ngIf="_prefix">
-	<ng-content select="[daffPrefix]"></ng-content>
-</ng-container>
-<div class="daff-tab__label">
-	<ng-content></ng-content>
-</div>
