+++ libs/design/tabs/src/tabs/tab-label/tab-label.component.html
@@ -1,0 +1,6 @@
+<ng-template #label>
+	<ng-container *ngIf="_prefix">
+		<ng-content select="[daffPrefix]"></ng-content>
+	</ng-container>
+	<ng-content></ng-content>
+</ng-template>
