+++ libs/design/tabs/src/tabs/tab-label/tab-label.component.scss
@@ -1,0 +1,4 @@
+:host {
+	display: flex;
+	flex-wrap: wrap;
+}
