+++ libs/design/tabs/src/tabs/tab-label/tab-label.component.spec.ts
@@ -1,0 +1,8 @@
+// import { DaffTabLabelComponent } from './tab-label.component';
+
+// describe('DaffTabLabelComponent', () => {
+//   it('should create an instance', () => {
+//     const directive = new DaffTabLabelComponent();
+//     expect(directive).toBeTruthy();
+//   });
+// });
