--- libs/design/tabs/src/tabs/tab-label/tab-label.directive.spec.ts
@@ -1,8 +1,0 @@
-import { DaffTabLabelDirective } from './tab-label.directive';
-
-describe('DaffTabLabelDirective', () => {
-  it('should create an instance', () => {
-    const directive = new DaffTabLabelDirective();
-    expect(directive).toBeTruthy();
-  });
-});
