+++ libs/design/tabs/src/tabs/tab-panel/tab-panel.component.scss
@@ -1,5 +1,5 @@
-.daff-tab {
-	$root: '.daff-tab';
+:host(.daff-tab-panel) {
+	$root: '.daff-tab-panel';
 	display: block;
 	padding: 1rem;
 }
