--- libs/design/tabs/src/tabs/tab/tab.component.scss
@@ -1,5 +1,0 @@
-.daff-tab-panel {
-	$root: '.daff-tab-panel';
-	display: block;
-	padding: 1rem;
-}
